Move test_pad into encoder_test
